Let RETURNING be the last stmt in the INSERT rule
In the psql console, even if the RETURNING is not last, an INSERT
RETURNING query on the view correctly returns the requested value,
when invoked via the PG library and AR with prepared statements,
no value is returned.

So, the RETURNING has been moved last, returning data from history
that'll contain the same data anyway. If the queries were to be
inverted, concurrency issues would have arisen and convolution as
the sequence would have need to be queried.