Initial version of the Time Machine
Unluckily, schema switching is not practical in an ActiveRecord::Relation
context, as its methods use both the model's find_by_sql and the adapter's
select_value, select_one, etc.

Wrapping is possible, e.g. by overriding singleton methods on each relation
instance, but it's simply not practical as many (private ones) of them have
to be overridden for each bit of functionality to work.

So, a simpler Arel::Relation#from approach has been used, that is way more
composable and doesn't rely on the inner workings of ActiveRecord::Relation.