Implement temporal associations between Chrono Models
This is the most hacky part of the library, as it is overriding the
ActiveRecord::Association class and the PostgreSQL ARel visitor.

The first is needed to alter the way association scopes are built,
to fetch data from the history if the associated record comes from
history; the second actually fixes ARel WITH visitor that currently
generates bad SQL.

The way history is queried has been changed using WITH (Common Table
Expressions) that allows table aliasing and prepending the temporal
constraints before the actual associated query.

Please note that .joins() and .includes() still do not work.