Partially revert e8a6ec3, as it fails with data structure changes
The triggers implementation introduced with e8a6ec3 were relying
on the history table colums order, as they were using NEW.*, then
the validity timestamps.

However, when new columns are added to the parent table, they end
up in the history table AFTER the validity timestamps, making the
triggers fail.

As currently on PG there is no way to alter a table columns order:

  http://wiki.postgresql.org/wiki/Alter_column_position

we are reverting to the previous behaviour, with explicit column
names in the triggers.