Remove migrations TableCache (2b8aea15). Fixes #56
TableCache is a premature optimisation of a method that is run only
in migrations context, and that holds a per-process cache of tables
and their temporal/non temporal status.

The cache was populated lazily, and maintained up to date with the
database status after running data definition methods (create_table
and so on).

In #56 use case, migrations are re-run in the same process instance
and the cache gets confused, as it assumes a table is still temporal
even though the database has been dropped and recreated.

As this complexity is to gain a little performance during database
migrations, it is now removed.