Replace dependant with dependent
"Dependant" is used as a noun when referring to someone who relies on
another person or source for support or care.

"Dependent" is used as an adjective to describe something that relies on
or is influenced by another thing.

It appears that `dependant` occurrences were always used as an adjective