Test against multiple PostgreSQL versions
This commit introduces an improvement in our testing strategy to ensure
broader compatibility across various PostgreSQL versions.

The primary focus remains on the latest stable release of PostgreSQL,
acknowledging the significance of maintaining support for older versions
to accommodate the longevity of applications.

To facilitate this, tests have been added to verify the application's
functionality against PostgreSQL 12.

Since `setup-postgres` action does not allow to create multiple
databases, consider a refactoring if ankane/setup-postgres#9 will be
merged.